requirements.txt
anyio==4.12.1
asyncio==4.0.0
certifi==2026.1.4
charset-normalizer==3.4.4
construct==2.10.70
construct-typing==0.7.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
jsonalias==0.1.1
python-telegram-bot==20.7
requests==2.32.5
sniffio==1.3.1
solana==0.36.11
solders==0.27.1
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
